+++ xbmc/cores/RetroPlayer/buffers/CMakeLists.txt
@@ -10,4 +10,11 @@ set(HEADERS BaseRenderBuffer.h
             RenderBufferManager.h
 )
 
+if(GBM_HAS_BO_MAP)
+  list(APPEND SOURCES RenderBufferGBM.cpp
+                      RenderBufferPoolGBM.cpp)
+  list(APPEND HEADERS RenderBufferGBM.h
+                      RenderBufferPoolGBM.h)
+endif()
+
 core_add_library(rp-buffers)
