+++ xbmc/cores/RetroPlayer/buffers/RenderBufferGBM.h
